Generalized a single KDDCup2009Dataset class + 3 task specific subclasses
