+++ ludwig/datasets/kdd_churn/config.yaml
@@ -5,8 +5,4 @@ download_urls:
   - "https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train_churn.labels"
   - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/churn/stratified_train_idx_churn.txt"
   - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/churn/stratified_test_idx_churn.txt"
-split_filenames:
-  train_file: orange_small_train.data.zip
-  test_file: orange_small_test.data.zip
-download_file_type: zip
 csv_filename: kdd_churn.csv
