+++ ludwig/datasets/kdd_upselling/config.yaml
@@ -5,8 +5,4 @@ download_urls:
   - "https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train_upselling.labels"
   - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/upselling/stratified_train_idx_upselling.txt"
   - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/upselling/stratified_test_idx_upselling.txt"
-split_filenames:
-  train_file: orange_small_train.data.zip
-  test_file: orange_small_test.data.zip
-download_file_type: zip
 csv_filename: kdd_upselling.csv
